int to float  tp_count

+++ models.py
@@ -165,16 +165,12 @@ class Client(BaseModel):
 class CalculatorStats(BaseModel):
     currency: str
     profit: float
-    tp_count: int
+    tp_count: float
     sl_count: float
     all_stats_count: int
     saved_stats_count: int
     max_profit: float
     min_loss: float
-
-# class Client(BaseModel):
-#     user: UserInfo | None = None
-#     subscribes: Optional[list[Subscribe], Subscribe, None] = None
 
 
 class Worker(BaseModel):
